Factor progress icon out into a new component to be used by the legend and data tables, use these icons in the LessonDataCell
