+++ apps/src/templates/sectionProgressV2/LegendItem.jsx
@@ -3,9 +3,7 @@ import PropTypes from 'prop-types';
 import styles from './section-progress-refresh.module.scss';
 import {BodyThreeText} from '@cdo/apps/componentLibrary/typography';
 import color from '@cdo/apps/util/color';
-import FontAwesome from '../FontAwesome';
-import ProgressBox from '../sectionProgress/ProgressBox';
-import classNames from 'classnames';
+import ProgressIcon from './ProgressIcon';
 
 export const ITEM_TYPE = Object.freeze({
   NOT_STARTED: 1,
@@ -21,50 +19,12 @@ export const ITEM_TYPE = Object.freeze({
   VALIDATED: ['circle-check', color.product_affirmative_default],
 });
 
+export const ITEM_TYPE_SHAPE = PropTypes.oneOf(Object.values(ITEM_TYPE));
+
 export default function LegendItem({itemType, labelText}) {
-  const needsFeedbackTriangle = () => (
-    <div className={classNames(styles.needsFeedback, styles.cornerBox)} />
-  );
-
-  const feedbackGivenTriangle = () => (
-    <div className={classNames(styles.feedbackGiven, styles.cornerBox)} />
-  );
-
-  const notStartedBox = () => (
-    <ProgressBox
-      started={false}
-      incomplete={20}
-      imperfect={0}
-      perfect={0}
-      lessonIsAllAssessment={false}
-    />
-  );
-
-  const viewedBox = () => (
-    <ProgressBox
-      started={false}
-      incomplete={20}
-      imperfect={0}
-      perfect={0}
-      lessonIsAllAssessment={false}
-      viewed={true}
-    />
-  );
-
   return (
     <div className={styles.legendItem}>
-      {itemType?.length && (
-        <FontAwesome
-          id={'uitest-' + itemType[0]}
-          icon={itemType[0]}
-          style={{color: itemType[1]}}
-          className={styles.fontAwesomeIcon}
-        />
-      )}
-      {itemType === ITEM_TYPE.NOT_STARTED && notStartedBox()}
-      {itemType === ITEM_TYPE.VIEWED && viewedBox()}
-      {itemType === ITEM_TYPE.NEEDS_FEEDBACK && needsFeedbackTriangle()}
-      {itemType === ITEM_TYPE.FEEDBACK_GIVEN && feedbackGivenTriangle()}
+      <ProgressIcon itemType={itemType} />
       <BodyThreeText className={styles.labelText}>{labelText}</BodyThreeText>
     </div>
   );
@@ -72,5 +32,5 @@ export default function LegendItem({itemType, labelText}) {
 
 LegendItem.propTypes = {
   labelText: PropTypes.string,
-  itemType: PropTypes.oneOf(Object.values(ITEM_TYPE)),
+  itemType: ITEM_TYPE_SHAPE,
 };
