+++ apps/src/templates/sectionProgressV2/LessonDataCell.jsx
@@ -4,21 +4,19 @@ import {studentLessonProgressType} from '../progress/progressTypes';
 import classNames from 'classnames';
 import styles from './progress-table-v2.module.scss';
 import {lessonHasLevels} from '../progress/progressHelpers';
+import {ITEM_TYPE} from './LegendItem';
+import ProgressIcon from './ProgressIcon';
 
 export default function LessonDataCell({lesson, studentLessonProgress}) {
-  const lessonData = React.useMemo(() => {
-    if (!lessonHasLevels(lesson)) {
-      return 'nolev';
-    }
-    if (!studentLessonProgress) {
-      return;
-    }
-    return Math.round(studentLessonProgress.completedPercent) + '%';
-  }, [lesson, studentLessonProgress]);
+  const noLevels = !lessonHasLevels(lesson);
+  const finished = studentLessonProgress?.completedPercent === 100;
+  const partiallyComplete = studentLessonProgress && !finished;
 
   return (
     <div className={classNames(styles.gridBox, styles.gridBoxLesson)}>
-      {lessonData}
+      {finished && <ProgressIcon itemType={ITEM_TYPE.SUBMITTED} />}
+      {partiallyComplete && <ProgressIcon itemType={ITEM_TYPE.IN_PROGRESS} />}
+      {noLevels && <ProgressIcon itemType={ITEM_TYPE.NO_ONLINE_WORK} />}
     </div>
   );
 }
